requirements.txt
fastapi
uvicorn
ollama
faiss-cpu
numpy
python-dotenv
pydantic